Normalizes in-game coordinate system.
Flips the y-axis to be positive in the usual Cartesian direction
(upward), rather than the direction that is common in computer
graphics (downward). Also changes angles to start at 0 radians
in the +x direction, changing to PI/2 in the +y direction,
etc, in the usual mathematical fashion. These changes are to
allow sin() and cos() to have the expected values for a given
angle, so that it is easier to reason about the space while
writing algorithms as a user.