Adds initial explosion animations.
When the player loses all health, an explosion is
initiated. The player is set to dead when the explosion
is finished, and that is when the game ends. This might
be problematic, because the game could still be won with
the player half-exploded, so it will need to be revisited.
DestructibleTargets also explode when they run out of health,
and are only marked as complete after the explosion has
finished. This also might result in some weird race
conditions, but it is good enough for an initial result.