Fixes explosions with flag for updating after game over.
Each game object has a flag updateOnGameOver, which by
default is set to false. If the flag is set to true (currently
only for explosion objects), the game will continue to update
those objects after the player has won or lost. All other objects
(with the flag set to false) will appear paused. When there are
no more objects that have the updateOnGameOver flag set to true
and the level is complete, the Level.doneUpdating() becomes true
and the update/draw loop is suspended as usual.