Adds a togglable debug mode.
The debug mode can be toggled with the 'x' key. When debug mode
is turned on, all objects are highlighted and all tooltips are shown.
In addition, a 100px grid is drawn behind objects but in front of
the star field.

I have edited the tooltip drawing code to reset the rendering context's
transform before drawing so that it always uses 16px fonts, instead
of (16 / scale)px fonts, which makes zooming less jittery for tooltips.