Reduces number of stars to improve performance.
Previously, when fully zoomed out, 5000+ particles would be rendered
for the star field, which could take more than 16 milliseconds (which is
more than the length of an entire frame). It would take significantly longer
on a mobile device. I have reduced the number of particles by a factor 10x,
which has reduced the time taken to render them by an order of magnitude.
All the stars now have the same radius, which was the maximum radius in
the previous scheme.

While this (for the most part) eliminates lag, the new star field kind of
looks like shit when compared side by side with the old one. I think I would
have to leverage WebGL in order to draw the old star field fast enough, but
it is not supported in all browsers (and I would have to write a compatibility
layer between Canvas2D and WebGL calls). If I can think of a way to render
the old star field faster, I will bring back the cool look.

Ways of rendering the star field faster that do NOT work:
* Drawing one circle to a separate canvas and blitting it to the main canvas
  instead of drawing a circle for each particle. When you zoom in, the bitmap
  needs to be scaled to elminate pixellation, which comes along with its
  own performance problems. In addition it only speeds up rendering by a
  factor of 2x.
* Tiling a big bitmap for the background. When the user starts moving fast
  enough, the repeating bitmap makes it look like the ship slows down, then
  starts moving backward, then slows down, then starts moving forward... I
  would have to implement a speed limit of less than min(canvas width, height)
  to avoid this, and that is not part of the game mechanics.