Adds asteroid prerendering.
Asteroids are rendered onto an off-screen canvas and then drawn
onto the canvas with drawImage(). If many asteroids are being
drawn, asteroid geometries of the same radius are cached so
that new asteroids created with the same radius do not need
to generate a geometry or render it to a canvas.