Passes only diff of world to worker thread.
The new world's array of game objects is diffed against the old world's
array of game objects. Changes (at entire-object granularity) are
sent to the thread instead, reducing the send-to-worker bottleneck
for large static levels. For large dynamic levels, this will decrease
performance as the diff takes a non-trivial amount of time to complete.